+++ libkbfs/bare_root_metadata_common.go
@@ -6,14 +6,6 @@
 
 package libkbfs
 
-import (
-	"context"
-	"time"
-
-	"github.com/keybase/kbfs/kbfscodec"
-	"github.com/keybase/kbfs/kbfscrypto"
-)
-
 // MakeInitialBareRootMetadata creates a new MutableBareRootMetadata
 // instance of the given MetadataVer with revision
 // MetadataRevisionInitial, and the given TlfID and
@@ -34,38 +26,4 @@ func MakeInitialBareRootMetadata(
 	}
 
 	return MakeInitialBareRootMetadataV3(tlfID, h)
-}
-
-// SignBareRootMetadata signs the given BareRootMetadata and returns a
-// *RootMetadataSigned object.
-func SignBareRootMetadata(
-	ctx context.Context, codec kbfscodec.Codec, signer kbfscrypto.Signer,
-	brmd BareRootMetadata, untrustedServerTimestamp time.Time) (
-	*RootMetadataSigned, error) {
-	// encode the root metadata and sign it
-	buf, err := codec.Encode(brmd)
-	if err != nil {
-		return nil, err
-	}
-
-	// Sign normally using the local device private key
-	sigInfo, err := signer.Sign(ctx, buf)
-	if err != nil {
-		return nil, err
-	}
-	var writerSigInfo kbfscrypto.SignatureInfo
-	if mdv2, ok := brmd.(*BareRootMetadataV2); ok {
-		writerSigInfo = mdv2.WriterMetadataSigInfo
-	} else {
-		buf, err = brmd.GetSerializedWriterMetadata(codec)
-		if err != nil {
-			return nil, err
-		}
-		writerSigInfo, err = signer.Sign(ctx, buf)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return MakeRootMetadataSigned(
-		sigInfo, writerSigInfo, brmd, untrustedServerTimestamp)
 }
